Fix comparison and narrowing errors reported by GCC

Commit migrated from https://github.com/dotnet/coreclr/commit/fd448f7d0c9d5ed3beee3b0ef078f29d97f3eefa

+++ src/coreclr/src/unwinder/amd64/unwinder_amd64.cpp
@@ -134,7 +134,7 @@ public:
     UCHAR operator[](int index)
     {
         int realIndex = m_offset + index;
-        UNWINDER_ASSERT(realIndex < sizeof(m_buffer));
+        UNWINDER_ASSERT(realIndex < (int)sizeof(m_buffer));
         return m_buffer[realIndex];
     }
 };
